submissions/terminalcli/requirements.txt
requests
whois
socialscan
dnspython
python-dotenv
